backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database (Tortoise ORM)
tortoise-orm[asyncpg]>=0.20.0
aerich>=0.7.1

# Async & Workers
celery>=5.3.0
redis>=5.0.0

# Image Processing & OCR
Pillow>=10.0.0
pytesseract>=0.3.10

# Security
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
argon2-cffi>=21.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0,<0.28.0
watchdog>=3.0.0
pytest>=7.4.0
